+++ Lib/test/test_setcomps.py
@@ -2,6 +2,8 @@ import doctest
 import traceback
 import unittest
 
+from test.support import BrokenIter
+
 
 doctests = """
 ########### Tests mostly copied from test_listcomps.py ############
@@ -154,33 +156,20 @@ class SetComprehensionTest(unittest.TestCase):
         # The location of an exception raised from __init__ or
         # __next__ should should be the iterator expression
 
-        class Iter:
-            def __init__(self, init_raises=False, next_raises=False):
-                if init_raises:
-                    1/0
-                self.next_raises = next_raises
-
-            def __next__(self):
-                if self.next_raises:
-                    1/0
-
-            def __iter__(self):
-                return self
-
         def init_raises():
             try:
-                {x for x in Iter(init_raises=True)}
+                {x for x in BrokenIter(init_raises=True)}
             except Exception as e:
                 return e
 
         def next_raises():
             try:
-                {x for x in Iter(next_raises=True)}
+                {x for x in BrokenIter(next_raises=True)}
             except Exception as e:
                 return e
 
-        for func, expected in [(init_raises, "Iter(init_raises=True)"),
-                               (next_raises, "Iter(next_raises=True)"),
+        for func, expected in [(init_raises, "BrokenIter(init_raises=True)"),
+                               (next_raises, "BrokenIter(next_raises=True)"),
                               ]:
             with self.subTest(func):
                 exc = func()
